Fixed more race conditions in install.
The same race seen in the eirini deployment and fixed with PR #46 is also present in `gitea`, `quarks`, and `traefik`. The same fix has been applied now (wait for pod existence first before waiting for it to be running).

All deployments waiting for more than one pod were changed to use an explicit list of pods to check and wait for.
They were also changed to use a loop over the pod names, if the set of pods contained more than one entry.

:warning: Attention: The tekton deployment code is different enough to be unsure if it has similar races or not, and the above fix cannot be applied as is. Some trials with pushing an app after install ran into claim of not finding tekton-related things (triggertemplate and triggerpipeline) while these looked to be present. Simply waiting a few minutes between `install` and `push` seemed to resolve these, indicating that there is indeed some kind of race present, albeit not exactly the same handled here, as it does not seem to involve pods.