Refactored kubernetes.Installer and paas.InstallClient.
  - Renamed Installer, now DeploymentSet
      - Changed interfaces to take a *Cluster.
  - Moved installation related methods (ShowOptions, Install)
    to the InstallClient.
      - Disabled their tests in kubernetes.

UNDO of attaching the `ui` to the deployment structures.
The `ui` is not inherent in the deployment set, nor deployment.
It is external and changing.
My conclusion is thatit should not be attached.
Like cluster is not attached. Same category of thing.

I can see aggregating the `Cluster` and `UI` arguments into a single structure.
Which we have, it is the `InstallClient`.
Going that route however requires movement of all the deployment related definition from `kubernetes` to `paas`, to avoid circular imports.