Remove Eirini first and (really) wait until the workloads ns is removed
The problem is that there is a Tekton Custom Resource evenlistener)
created in the eirini-workloads namespace. That has a finalizer handled
by a controller deployed as part of the Tekton deployment. If the Tekton
deployment is removed first, it happens that the controller is removed
before the CRD so when the time comes to remove the Custom Resource,
there is no controller to remove the finalizer thus the Custom Resource
stays there preventing the eirini-workloads namespaces from being
deleted.

See also: https://book.kubebuilder.io/reference/using-finalizers.html