Set the traefik externalIPs when we install traefik.
In the case of k3d/k3s this IP is already handled by k3s but in the
other cases we need to set it to something to be able to talk to the
cluster. In the future, when more clusters are supported, we may have
situations in which we install traefik and we need to wait until a
public IP is assigned by the loadbalancer. This is not yet the case.