feat: Extended ui package with verbosity levels. feat: Extended ui package with methods helping with keeping the output looking good regardless of level (newline control). fix:  Bumped the level of various progress messages (not the dots) to hide them at default verbosity. fix:  Tweaked the messaging in `push` for the different verbosity levels. fix:  Tweaked the messaging before various progress displays to be conditional on verbosity.
The last fixes ensure that the dots of the dot-progress look nicely
regardless of verbosity.  At standard verbosity, when the progress
messages are hidden the dots attach to the last message before the
progress indicator. With progress display active at higher verbosity
the dots then attach to the individual progress messages.