fix: Use readiness probes to check if a Pod is "running"
Note: For pods that end in "Completed" state this check won't work. But
we shoulnd't care about such pods and if we ever do, we should implement
another method. This one is about "running" pods.

because checking the individual container statuses doesn't mean the pod
is ready to server requests.

https://kubernetes.io/docs/concepts/workloads/pods/pod-lifecycle/#pod-conditions