feat: Determine bound services, per app. Show when listing apps
Extended the application backend with function to determine the set of
bound services. Further extended the `apps` cli command to show the
bound services of each app.

feat: Unbind bound services as part of deleting an application

This also includes moving the implementation of the basic deletion
into the application package.

feat: Invert services of app to apps of services, and show

Extended services table with a column to show the apps using that
service. Added backend function getting the services per app, for all
apps, and inverting that into apps per service, for all services.

fix: Prevent deletion of service bound to application(s)

chore: Added comments to various public functions.