feat: Allow the usage of non-embbeded files when developing
because this way it's easier to change html/javascript/css without the
need to compile the binary again. In combination with the predefined
--port flag, it should make the front-end development a lot easier
through the browser.

run:

LOCAL_FILESYSTEM=true carrer gui --port 8080

and then open the browser at localhost:8080 . Change some file inside
the embedded-web-files directory and reload the page in the browser and voila!