refactor: Remove 'gui' command and refactor packages
we don't need the gui, a browser should be enough. If the user wants a
launcher they can create a wrapper script easily.

Also simplify how we handle LOCAL_FILESYSTEM for development, no need to
use viper for it. We read it directly from os.Getenv.

Finally implement a logging middleware for requests.