refactor: create method to start the server in a go routine
On the upcoming refactoring we are going to start the server on a local
port when the user runs any cli command. E.g. when listing apps, we
will:

- start the server on a local port
- make a request to the server
- present the results
- stop the server

When all cli commands are implemented through the api, we can move the
server inside the cluster. Then instead of starting a local server we
will simply target the remote one (through a flag/config).