fix: Don't allow triggering of the staging pipeline without a password
As part of the research for mTLS, we discovered that it's possible to
trigger the staging of an application from another application's
container. More here:

https://github.com/SUSE/carrier/issues/130#issuecomment-813314112

This PR closes this hole but using a CEL interceptor
(https://github.com/tektoncd/triggers/blob/v0.11.1/docs/eventlisteners.md#cel-interceptors).

We still don't verify who the caller is, like mTLS would do but it's
better than nothing since we control who has access to that secret.