fix: Ensure carrier-server is restarted every time we patch
because the second time the patching will be a no-op and will not force
a restart of the server although the binary might have changed.

Also use rollout status to make sure the server is up before we return.
This will make tests wait until the server is up before they start testing.