fix: Login to dockerhub before running tests
because if a registry mirror is not running, the acceptance tests will
try to start it. If the image is not yet there (like it happens on new
github runners), docker command will try to pull it. On self-hosted
runners which share an external IP with lots of other machines, this
could immediately hit the rate limiting of dockerhub.