Fix race conditions that causes closing of a closed channel
we simply ignore the panic if the channel is closed already.
As far as I can tell, this can only happen if the Stream closes the ctx
at the same time as the pod disappears. The disappearing of the Pod
results in (t *Tail) Close() being called (which then closes t.closed on
line 172) and the end of the Stream results in ctx.Done() to be closed which also
results in t.closed being closed (line 159).

We solve this here by ignoring the closed channel if someone else closed
it already. Anything that should happen when the channel closes, must
have already happened when it was first closed. Closing it again would
have no effect anyway (even if it was possible).