Fix reloading of apps during repush
We use the previous git commit of the app storage to have an unique
identifier in the image url which changes every time the app is
repushed.

Even after the first push there are at least two commits because we have
an initial commit and one for the first push.

Even if the app is not changed, only pushed again the app will
retrigger, because our current implementation creates a new git commit
every time we push.

Fixes #262