feat: Enable the binding of services to an app as part of `push`
INCOMPATIBILITY: API CHANGE
	Modified service bind requests to carry an array of service
	names, instead of a single name.

Enables the binding of multiple services to one app in a single API call.

Modified the receiver to match.
Note that the receiver still performs the binds sequentially.
See implementation notes below.

Updated the `service bind` command to match.

Extended the `push` command to accept services to be bound, via a new
option `--bind`/`-b` for `push`. Note that the option allows the
specification of multiple services (1). The code removes duplicates in
the specified set of services.

Added command completion for `--bind` (service names).

Factored generation of many errors into helper functions.
Ensures more consistent messaging.

----

__Simple__ implementation: Internally does what the user would
otherwise do manually, i.e.

  - push the app
  - bind the specified services to the app, one after the other

The only optimization right now is that the entire set of services is
transfered to the server in a single bind request. See API change above.

__Optimizations__ like adding the services to the app kube deployment
as part of starting the app, instead of via separate updates, are
__defered__ until we have separated uploading the app sources from
starting the app (2).

----

  1. Example: `... --bind service1,service2 --bind service3`

  2. Currently app start is automatically triggered by the source
     upload (into gitea). See issues #347 and #358 for the requisite
     work needed.