Don't write .kube files in the app source repo
but put them as templates in the pipelines definition.

Among other things it also:

Fixes #394

Not having the .kube directory in the sources also allows us to
implement a git ops workflow in the future (where the user simply pushes to gitea
without using `epinio push`). This is a first step.

Co-authored-by: Mario Manno <mario.manno@suse.com>