Retry applying tekton resources if webhook fails during installation (fixes #404)
It seems the readiness probe in Tekton is not effective.
If we try to apply a Tekron resource immediately after the pod is ready,
the service will fail.

> `Post "https://tekton-pipelines-webhook.tekton.svc:443/defaulting?timeout=10s": dial tcp 10.43.110.8:443: connect: connection refused`

Tried workarounds:
* wait for deployment to be available
* wait for webhook-cert secret to appear and have the right keys
* remove HorizontalPodAutoscaler
* add a startup probe
* point probes to https ports
* change probe endpoint to admissionwebhooks path (`/defaulting`)

The only thing that worked, was waiting ~15s before applying the resources.

Might be related to: https://github.com/tektoncd/pipeline/issues/2207