Add new endpoint for staging logs
because one could request staging logs for a new application. In that
case, there is no applicatino yet and we should solely rely on stagingID
to filter logs.

We use the same backend code to stream logs with the following logic:
if there is a stageID it means we want staging logs, if not we look for
  an app name and stream logs for it.