Use app reference instead of app model
Removes the app model from the workload struct.
It was there to maintain compatibility with existing code, but it turns
out the workload manager doesn't need access to the app model.

We pass just the reference to the app (org, name). The code is no longer
able to mutate the passed in app model.
The one function that did need to mutate the model (Complete()) always
returned it.