fix: Drop general silencing of the usage display on cli command line errors
Dynamically silence usage display for errors unrelated to the
commandline, i.e. from the backend and such. In most places this
simply means silencing the moment we are in the command function. In
very few places this silence has to be lifted again for specific
issues.

The ensemble command specs are unchanged.