fix: attempt to retrieve stage id of an app without workload
As in, don't.

Dropped the mix of querying pods and deployment. Looking at the
deployment is enough to get all the info we need (status and stageId).

This efectively inlined `DeploymentStatus` into its single caller, and
extended it. Thus the removal of the function itself.

New tests check that scaling an app down to zero instances keeps
`list` and `show` working.