Cleanup all pipelineurns after updating the deployment
because that's the way it worked before. The reason is that the staging
endpoint shouldn't delete old staging logs until the new artifact is
actually used (deployed). Someone may try to fetch the staging logs
between a new staging process has completed and the deployment not being
updated yet. Given the deploy endpoint is a separate one now, it means
one may never even deploy the new artifact. The staging logs of the
running image should always be available.

This commit moves the cleanup to the point after the new image is
actually deployed.

Co-authored-by: Dimitris Karakasilis <DKarakasilis@suse.com>