Fix golangci-lint workflow
The workflow didn't work previously, because the generated files were
missing.

This PR adds detection for schedule vs. pull_request events.
In schedule mode, we use a relaxed config, to avoid to many warnings.
In PR mode the config is stricter, so new code lints.
The PR mode config is also in the default config location and should be
picked up by IDEs.