chore: Further (internal) renames org/namespace feat: Extended namespace index endpoint feat: New and separate namespace match endpoint feat: Extended `namespace delete` command with command completion. fix: Env name match handler.
The extended `Index` endpoint delivers an array of namespace objects.
The objects carry not only the namespace name, but also arrays of
apps and services (names only) in the namespace.

namespace matching for command completion going through index endpoint
would now compute lots of superfluous data. The new endpoint does just
as the old index did, generate a list of namespace names.
New, it does the matching server side.

Note the double registration of the match endpoint with and without pattern.
The second catches invokation with empty string as the pattern!

Applied the double registration to the env match endpoint also, as fix.