Disable the tekton affinity-assistant
We get one affinity assistant per "use" of the workspace (not per PVC, not per
Task). We are mounting 2 subPaths of the same PVC in the same Task and that's why
we get 2 affinity assistants.

This is a problem when the provider allows you to use a PVC from 2 different nodes
but not in parallel. That's because your 2 affinity assistants could land on 2
different nodes and they would both try to mount the PVC, which is not allowed.

So, depending on the storage class attributes, this could fail.

Notes:

* https://github.com/tektoncd/pipeline/issues/3731
* our tasks run sequentially.
* previously resulted in `Multi-Attach error for volume "pvc-aac89874" Volume is already used by pod(s) affinity-assistant-38e-0`