Fix error when CRD doesn't have a conditions field yet
We got this error randomly in CI:

```
error: .status.conditions accessor error: <nil> is of the type <nil>, expected []interface{}
```

possibly originating here:

https://github.com/kubernetes/kubernetes/blob/master/staging/src/k8s.io/apimachinery/pkg/apis/meta/v1/unstructured/helpers.go#L137

This commit is re-using existing code to wait for CRDs. We no longer
check conditions but it works as is in existing code. If this ever
fails, we may need to extend the WaitForCRD function to also check the
conditions field for establishment.