Add an Error attribute to the app object
Listing commands and api endpoints use it now to avoid erroring out the
whole command. Only the app with the error will now show the error in
the view.

This can happen if while we are collecting the information for all the
apps, some app gets deleted by another process. This would cause the
whole request to fail but now it will allow the rest of the apps to be
printed with no error.