refactor: Removed servicesToApps. Replaced with collection of more specialized functions.
The new functions return app names and apps, as needed, for one or all namespaces, all or specific services.
Note that not all combinations implied in the above are actually needed and implemented.

The internals are also written to use the application-associated secrets storing the desired/bound services for an app. As Dimitris pointed out, we can use them to get the information faster than poking into the applications as we did before, and we can query them quicker also (filtered lists).

This should make querying bounds apps for services quicker to determine.