Make epinio server unaware of the force-kube-internal-registry-tls flag
This will allow the declarative installer to use the registry localhost trick.
Now the epinio server relies only on the existence of a localhost url in the
connection details secret to decide wether to rewrite the image url or
not.

It's up to the installer (any of them) to decide if a localhost url
should be in the secret. Our current installer uses the force-kube**
flag, the declarative installer can simply allow an array of registry
urls to be passed through helm values.