[PoC] Implement a channel based pub/sub system
TODO: Implenent a "subscribe" endpoint which will keep track of all
subscribers. A "subscribe" request will result in a websocket connection
which should be sent to the main "handler" (the new go routine running
alongside the server). That hanlder will make sure each message is sent
to each interested party.