Don't test the installation in the epinio acceptance test suite
Because we need Traefik before we install Epinio to make the chartmuseum
ingress to work. For that reason we let k3d install its own Traefik but
that one is not configured to redirect http to https.

In any case, testing how the various dependencies are installed is not
part of this test suite. It belongs to the installation test suites,
thus we remove the rest of the tests which weren't testing much anyway.