Make test fail (like it should)
by properly using the minio client pod. We used `kubectl run -it` which
in some cases simply printed errors like:

"if not prompt is displayed try hitting Enter" (or something like that)
or
"Error attaching, falling back to logs: Internal error occurred: error attaching to container:"

Now we start the pod and we run each command with `kubectl exec`. We
don't need to call `bash` anymore and we wait for the pod to be ready
before we run any commands. The output should always be the list of
blobs.

Now the test fails because the "stage" endpoint was never supposed to
delete any blobs. That's the "deploy" endpoints.

TODO: Move the tests to the deploy endpoint tests.