Prevent ApplicationUpdateRequest service binding changes during issues
Before service binding changes were accepted even though there was an
error for example with unknown service bindings.

If an unknown service binding was provided as part of
ApplicationUpdateRequest, it was filtered out but the changes were still
applied, which was an empty array of service bindings resulting in an
unintentional removal of all service bindings.