Use scratch as base image for epinio-server
scratch does not provide a readable write dir so we provide one via
volume mount.

This does have the advantage that we properly split the tmp dir from the
node.
The tmp dir is needed to pull an app from git.

The remaining penalty of scratch is that we can not shell into it
easily, but there are workaround for it which we can document.