Use locally built epinio image for ci
We always tag our latest released Docker image as `latest`. By using
latest for patching, it allows patching still to be used the same way as
before in non-k3d environments.
But at the same time when using the make target `prepare-environment-k3d`
the locally built docker image is used and this allows proper ci and
manual testing of the whole setup including changes in the Dockerfile,
not just our binary.