Ignore services that were deleted in the middle of an "index" request
E.g. This happens in tests when a ginkgo node is listing all services
from all namespaces while some other ginkgo node is deleting a service
in that list. Our "index" code, first fetchese the list of services and
then finds details for each one of them. If a service is gone between
listing and fetching the details, the old version would make the whole
request fails. Now the service is simply ignored.