Fix six dependabot alerts
All direct requirements were bumped to see if that would fix the issues.

containerd is added in so many different versions, even beta versions to
the go.sum that I think it is ok to replace it with a wild card version.

Dependabot should alert us if there is an issue with that version.