fix: make app logging closer goroutine on-demand and disable it for `app logs` command.
For `app logs` context cancellation is late.
Closer races against process termination
When closer wins the race a panic occurs, double close of the message channel

However, closer is not required, because websocket is closed by server!
Closing by user would be sigterm (^C).
That kills the process and goroutines with no time for this kind of finalization.