Introduce "events" package that handles the messaging functionality
This code can be tested using a receiver like this one:

https://github.com/rabbitmq/rabbitmq-tutorials/blob/master/go/receive.go

The rabbitmq service needs to be exposed locally with:

```
kubectl port-forward --namespace epinio svc/epinio-rabbitmq 5672:5672
```

and the credentials for the consumer are stored in a secret called
epinio-rabbitmq.

NOTE: Uninstalling epinio doesn't delete the rabbitmq PVC which will
result in authentication errors if you re-install Epinio without
deleting the PVC first.