fix: mishandling of commit ids in git upload, and other issues
As part of that also fixed the misused optimization `SingleBranch`.
Cannot be applied of there is no revision.
Ditto when we have to handle a commit id.

Generally restructured the code with helper function to make the higher-level flow visible.
Added design notes at the top explaining the reasoning behind that flow.

Thanks to Enrico for `repository.ResolveRevision(plumbing.Revision(revision))`.