fix use of wrong bound variable in calculation of app/config map
we have to handle the just found apps, not the collection of all apps found in previous iterations.
in the case of a single configuration/service this causes the map to be empty
this in turn causes partial deletion of the service.
the linkage to the app is not properly broken.
trying to bind something else to that app then fails trying to get data from the deleted service